Fix review comments and code refactor
