--- app/javascript/src/components/routes.tsx
@@ -1,74 +1,0 @@
-import ErrorPage from "common/Error";
-import { Roles } from "constants/index";
-import ClientDetails from "./Clients/Details";
-import ClientList from "./Clients/List";
-import EditInvoice from "./Invoices/Edit";
-import GenerateInvoices from "./Invoices/Generate";
-import Invoice from "./Invoices/Invoice";
-import InvoicesList  from "./Invoices/List";
-import Payments from "./payments";
-import ProjectDetails from "./Projects/Details";
-import ProjectList from "./Projects/List";
-import OutstandingInvoiceReport from "./Reports/outstandingInvoices";
-import ReportList from "./Reports/reportList";
-import RevenueByClientReport from "./Reports/revenueByClient";
-import TimeEntryReports from "./Reports/timeEntry";
-import TotalHoursReport from "./Reports/totalHoursLogged";
-import PlanSelection from "./Subscriptions/PlanSelection";
-import TimeTracking from "./time-tracking/Index";
-
-const ClientsRoutes = [
-  { path: "" ,Component: ClientList },
-  { path: ":clientId" ,Component: ClientDetails }
-];
-
-const ReportsRoutes = [
-  { path: "" ,Component: ReportList },
-  { path: "time-entry" ,Component: TimeEntryReports },
-  { path: "revenue-by-client" ,Component: RevenueByClientReport },
-  { path: "outstanding-overdue-invoice" ,Component: OutstandingInvoiceReport },
-  { path: "total-hours" ,Component: TotalHoursReport }
-];
-
-const InvoicesRoutes = [
-  { path: "" , Component: InvoicesList },
-  { path: "generate" ,Component: GenerateInvoices },
-  { path: ":id/edit" ,Component: EditInvoice },
-  { path: ":id" ,Component: Invoice },
-  { path: "*" ,Component: ErrorPage }
-];
-
-const ProjectsRoutes = [
-  { path: "" ,Component: ProjectList },
-  { path: ":projectId" ,Component: ProjectDetails },
-  { path: "*" ,Component: ErrorPage }
-];
-
-const SubscriptionsRoutes = [
-  { path: "" ,Component: PlanSelection },
-  { path: "*" ,Component: ErrorPage }
-];
-
-const PaymentsRoutes = [
-  { path: "" ,Component: Payments },
-  { path: "*" ,Component: ErrorPage }
-];
-
-const TimeTrackingRoutes = [
-  { path: "" ,Component: TimeTracking },
-  { path: "*" ,Component: ErrorPage }
-];
-
-const { ADMIN, OWNER, BOOK_KEEPER, EMPLOYEE } = Roles;
-
-const ROUTES = [
-  { path: "clients" ,subRoutes: ClientsRoutes, authorisedRoles: [ADMIN, OWNER, EMPLOYEE] },
-  { path: "invoices" ,subRoutes: InvoicesRoutes, authorisedRoles: [ADMIN, OWNER, BOOK_KEEPER] },
-  { path: "reports" ,subRoutes: ReportsRoutes, authorisedRoles: [ADMIN, OWNER] },
-  { path: "projects" ,subRoutes: ProjectsRoutes, authorisedRoles: [ADMIN, OWNER, EMPLOYEE] },
-  { path: "subscriptions" ,subRoutes: SubscriptionsRoutes, authorisedRoles: [ADMIN, OWNER] },
-  { path: "payments" ,subRoutes: PaymentsRoutes, authorisedRoles: [ADMIN, OWNER, BOOK_KEEPER] }
-  ,{ path: "time-tracking" ,subRoutes: TimeTrackingRoutes, authorisedRoles: [ADMIN, OWNER, EMPLOYEE, BOOK_KEEPER] }
-];
-
-export default ROUTES;
