+++ app/javascript/src/constants/index.tsx
@@ -24,6 +24,17 @@ export enum Roles {
   EMPLOYEE = "employee",
 }
 
+export enum Paths {
+  SIGN_IN = "/user/sign_in",
+  CLIENTS = "/clients",
+  INVOICES = "/invoices",
+  REPORTS = "/reports",
+  PROJECTS = "/projects",
+  SUBSCRIPTIONS = "/subscriptions",
+  PAYMENTS = "/payments",
+  TIME_TRACKING = "/time-tracking",
+}
+
 export const TOASTER_DURATION = 3000;
 
 export const getToasterIcon = (type) => {
